requirements.txt
altgraph==0.10.2
amqp==1.4.9
anyjson==0.3.3
appdirs==1.4.3
backports-abc==0.5
bdist-mpkg==0.5.0
BeautifulSoup==3.2.1
beautifulsoup4==4.6.0
billiard==3.3.0.23
boto==2.46.1
bz2file==0.98
celery==3.1.18
certifi==2017.4.17
Django==1.11
gensim==2.0.0
json2html==1.1.1
kombu==3.0.37
lxml==3.7.3
macholib==1.5.1
matplotlib==1.3.1
modulegraph==0.10.4
mygene==3.0.0
MySQL-python==1.2.5
mysqlclient==1.3.10
myvariant==0.3.1
nltk==3.2.2
nose==1.3.7
numpy==1.12.1
packaging==16.8
pandas==0.19.2
Pattern==2.6
py2app==0.7.3
pyparsing==2.2.0
python-dateutil==1.5
python-Levenshtein==0.12.0
pytz==2013.7
redis==2.10.3
requests==2.13.0
scipy==0.19.0
singledispatch==3.4.0.3
six==1.10.0
smart-open==1.5.1
tornado==4.5.1
virtualenv==15.1.0
xattr==0.6.4
zope.interface==4.1.1
xlrd==0.9.3
langdetect==1.0.7
urllib2
Google-Search-API
bs4==0.0.1
html5lib
jieba
djangorestframework